+++ finagle-http2/src/test/scala/com/twitter/finagle/http2/transport/client/ClientSessionImplTest.scala
@@ -2,17 +2,22 @@ package com.twitter.finagle.http2.transport.client
 
 import com.twitter.conversions.DurationOps._
 import com.twitter.finagle.http.Request
+import com.twitter.finagle.http2.param.MaxConcurrentStreams
+import com.twitter.finagle.http2.param.MaxRequestsPerSession
 import com.twitter.finagle.netty4.http.Bijections
-import com.twitter.finagle.{FailureFlags, Stack, Status}
-import com.twitter.util.{Await, Awaitable}
+import com.twitter.finagle.FailureFlags
+import com.twitter.finagle.Stack
+import com.twitter.finagle.Status
+import com.twitter.util.Await
+import com.twitter.util.Awaitable
 import io.netty.buffer.Unpooled
 import io.netty.channel.embedded.EmbeddedChannel
-import io.netty.channel.{Channel, ChannelHandler, ChannelInitializer}
-import io.netty.handler.codec.http2.{
-  Http2MultiplexCodec,
-  Http2MultiplexCodecBuilder,
-  Http2StreamFrameToHttpObjectCodec
-}
+import io.netty.channel.Channel
+import io.netty.channel.ChannelHandler
+import io.netty.channel.ChannelInitializer
+import io.netty.handler.codec.http2.Http2MultiplexCodec
+import io.netty.handler.codec.http2.Http2MultiplexCodecBuilder
+import io.netty.handler.codec.http2.Http2StreamFrameToHttpObjectCodec
 import org.scalatest.funsuite.AnyFunSuite
 
 class ClientSessionImplTest extends AnyFunSuite {
@@ -111,9 +116,14 @@ class ClientSessionImplTest extends AnyFunSuite {
     }
   }
 
-  test("Status is Closed when we're less than 50 streams away from exhausting the identifiers") {
+  test(
+    "By default status is Closed when we're less than 50 streams away from exhausting the identifiers") {
     new Ctx {
       assert(clientSession.status == Status.Open)
+
+      multiplexCodec.connection.local.createStream(Int.MaxValue - 100, false)
+      assert(clientSession.status == Status.Open)
+
       // client streams are odd streams so to be less than 50 we need to multiply by 2.
       multiplexCodec.connection.local.createStream(Int.MaxValue - 100 + 2, false)
       assert(clientSession.status == Status.Closed)
@@ -128,6 +138,23 @@ class ClientSessionImplTest extends AnyFunSuite {
       assert(clientSession.status == Status.Open)
       multiplexCodec.connection.local.createStream(1, false)
       assert(clientSession.status == Status.Busy)
+    }
+  }
+
+  test("Status is closed when we have exhausted the max requests per session limit") {
+    new Ctx {
+      override def params: Stack.Params = Stack.Params.empty
+        .+(MaxConcurrentStreams(Option(128L))).+(
+          MaxRequestsPerSession(Option(50000L))
+        )
+      assert(clientSession.status == Status.Open)
+
+      private val expectedHighWatermark: Int = 50000 * 2 - 128 * 2
+      multiplexCodec.connection.local.createStream(expectedHighWatermark - 1, false)
+      assert(clientSession.status == Status.Open)
+
+      multiplexCodec.connection.local.createStream(expectedHighWatermark + 1, false)
+      assert(clientSession.status == Status.Closed)
     }
   }